+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 //todo
 ## [v2.4.5](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.4.5)
 ### Added
-- Handle credential serialization error.
+- Handle credential serialization error and JWT parsing error.
 
 ## [v2.4.3](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.4.3)
 ### Changed